fix softlock bug
a softlock would occur when the last person to choose their defending
location drops or is kicked.  This commit makes sure to check if the
player being removed had any units in the "choose defender location".
If so, to run the logic which checks if that location is then clear of
any units and triggers the game starting timer if so.